Refactoring Workload Details page
